Move sfen input/output streams to sfen_stream.h
